test(evaluators): consolidate dogfooding tests into existing files
Per the project preference for test minimalism, fold the new tests added
across D4–D7 into the existing test files for the same surfaces rather
than carrying parallel files per-changeset.

- Move D4/D5/D6 Playwright tests from dogfooding-fixes.spec.ts into the
  existing code-evaluators.spec.ts (D6 reuses prereq sandbox/dataset
  fixtures; D4 creates its own env-var sandbox; D5 stays in its own
  describe since it doesn't share state with the eval flows).
- Move D7 getDependencyPackages/getDependencyPreview and D4
  getDisplaySandboxConfig describes into the existing
  formValuesToConfigPatch.test.ts (same module under test: utils.tsx).
- Delete the three now-empty per-changeset test files.

Net: -134 lines, no coverage change.
