+++ app/src/pages/settings/sandboxes/__tests__/formValuesToConfigPatch.test.ts
@@ -1,7 +1,12 @@
 import { describe, expect, it } from "vitest";
 
 import type { BackendInfo, SandboxConfigFormValues } from "../types";
-import { formValuesToConfigPatch } from "../utils";
+import {
+  formValuesToConfigPatch,
+  getDependencyPackages,
+  getDependencyPreview,
+  getDisplaySandboxConfig,
+} from "../utils";
 
 type PartialBackend = Pick<
   BackendInfo,
@@ -116,5 +121,186 @@ describe("formValuesToConfigPatch — capability-only output", () => {
     );
 
     expect(result).not.toHaveProperty("internet_access");
+  });
+});
+
+describe("getDependencyPackages", () => {
+  it("trims whitespace and drops empty lines", () => {
+    expect(getDependencyPackages("  numpy  \n\n pandas\n")).toEqual([
+      "numpy",
+      "pandas",
+    ]);
+  });
+
+  it("returns empty array for empty input", () => {
+    expect(getDependencyPackages("")).toEqual([]);
+    expect(getDependencyPackages("   \n  ")).toEqual([]);
+  });
+});
+
+describe("getDependencyPreview", () => {
+  it("returns null when no language is advertised", () => {
+    expect(
+      getDependencyPreview({
+        packagesText: "numpy",
+        dependenciesLanguage: null,
+        backendType: "E2B",
+      })
+    ).toBeNull();
+    expect(
+      getDependencyPreview({
+        packagesText: "numpy",
+        dependenciesLanguage: undefined,
+        backendType: "E2B",
+      })
+    ).toBeNull();
+  });
+
+  it("returns null when packages list is empty", () => {
+    expect(
+      getDependencyPreview({
+        packagesText: "",
+        dependenciesLanguage: "PYTHON",
+        backendType: "E2B",
+      })
+    ).toBeNull();
+    expect(
+      getDependencyPreview({
+        packagesText: "   \n  ",
+        dependenciesLanguage: "PYTHON",
+        backendType: "E2B",
+      })
+    ).toBeNull();
+  });
+
+  it("renders pip install for Python on E2B", () => {
+    expect(
+      getDependencyPreview({
+        packagesText: "numpy\npandas==2.0",
+        dependenciesLanguage: "PYTHON",
+        backendType: "E2B",
+      })
+    ).toBe("pip install numpy pandas==2.0");
+  });
+
+  it("renders pip install for Python on Daytona", () => {
+    expect(
+      getDependencyPreview({
+        packagesText: "numpy",
+        dependenciesLanguage: "PYTHON",
+        backendType: "DAYTONA_PYTHON",
+      })
+    ).toBe("pip install numpy");
+  });
+
+  it("renders image.pip_install for Python on Modal", () => {
+    expect(
+      getDependencyPreview({
+        packagesText: "numpy\npandas==2.0",
+        dependenciesLanguage: "PYTHON",
+        backendType: "MODAL",
+      })
+    ).toBe('image.pip_install("numpy", "pandas==2.0")');
+  });
+
+  it("renders unavailable for TypeScript", () => {
+    expect(
+      getDependencyPreview({
+        packagesText: "lodash",
+        dependenciesLanguage: "TYPESCRIPT",
+        backendType: "DENO",
+      })
+    ).toBe("preview unavailable for typescript");
+  });
+});
+
+describe("getDisplaySandboxConfig", () => {
+  it("redacts literal env_var values", () => {
+    const result = getDisplaySandboxConfig({
+      env_vars: [
+        { kind: "literal", name: "OPENAI_API_KEY", value: "sk-secret" },
+      ],
+    });
+
+    expect(result).toEqual({
+      env_vars: [
+        {
+          kind: "literal",
+          name: "OPENAI_API_KEY",
+          value: "<redacted>",
+        },
+      ],
+    });
+  });
+
+  it("preserves secret_ref entries unchanged", () => {
+    const config = {
+      env_vars: [
+        {
+          kind: "secret_ref",
+          name: "OPENAI_API_KEY",
+          secret_key: "openai_key",
+        },
+      ],
+    };
+
+    expect(getDisplaySandboxConfig(config)).toEqual(config);
+  });
+
+  it("preserves siblings of env_vars", () => {
+    const result = getDisplaySandboxConfig({
+      env_vars: [{ kind: "literal", name: "A", value: "1" }],
+      internet_access: { mode: "allow" },
+      dependencies: { packages: ["numpy"] },
+    });
+
+    expect(result).toEqual({
+      env_vars: [{ kind: "literal", name: "A", value: "<redacted>" }],
+      internet_access: { mode: "allow" },
+      dependencies: { packages: ["numpy"] },
+    });
+  });
+
+  it("does not mutate the input object", () => {
+    const original = {
+      env_vars: [{ kind: "literal", name: "A", value: "1" }],
+    };
+    const snapshot = JSON.parse(JSON.stringify(original));
+
+    getDisplaySandboxConfig(original);
+
+    expect(original).toEqual(snapshot);
+  });
+
+  it("passes through non-object configs unchanged", () => {
+    expect(getDisplaySandboxConfig(null)).toBeNull();
+    expect(getDisplaySandboxConfig(undefined)).toBeUndefined();
+    expect(getDisplaySandboxConfig("string")).toBe("string");
+    expect(getDisplaySandboxConfig(42)).toBe(42);
+  });
+
+  it("passes through malformed env_vars (non-array) unchanged", () => {
+    const result = getDisplaySandboxConfig({
+      env_vars: "not-an-array",
+    });
+    expect(result).toEqual({ env_vars: "not-an-array" });
+  });
+
+  it("passes through malformed env_var entries (non-object) unchanged", () => {
+    const result = getDisplaySandboxConfig({
+      env_vars: ["string-entry", null, 42],
+    });
+    expect(result).toEqual({
+      env_vars: ["string-entry", null, 42],
+    });
+  });
+
+  it("does not throw on literal entry missing value field", () => {
+    const result = getDisplaySandboxConfig({
+      env_vars: [{ kind: "literal", name: "A" }],
+    });
+    expect(result).toEqual({
+      env_vars: [{ kind: "literal", name: "A" }],
+    });
   });
 });
