+++ app/tests/code-evaluators.spec.ts
@@ -209,6 +209,64 @@ async function expectEvaluatorDetailsPage(page: Page, evaluatorName: string) {
   ).toBeVisible();
 }
 
+async function createE2BSandboxWithLiteralEnvVar(
+  page: Page,
+  configName: string,
+  envVarName: string,
+  envVarValue: string
+) {
+  await page.goto("/settings/sandboxes");
+  await page.waitForURL("**/settings/sandboxes");
+
+  await page.getByRole("button", { name: "New Sandbox" }).click();
+  const dialog = page.getByRole("dialog");
+  await expect(dialog).toBeVisible();
+
+  await dialog.getByLabel("Provider").fill("E2B");
+  await page.getByRole("option", { name: /E2B/i }).first().click();
+
+  await expect(dialog.getByText("Environment Variables")).toBeVisible();
+  await dialog.getByRole("button", { name: "Add Variable" }).click();
+  await dialog.getByLabel("Name").last().fill(envVarName);
+  await dialog.getByLabel("Value").last().fill(envVarValue);
+
+  await dialog.getByLabel("Name").first().fill(configName);
+
+  await Promise.all([
+    page.waitForResponse((resp) =>
+      isGraphQLMutationResponse(
+        resp,
+        "SandboxConfigDialogCreateSandboxConfigMutation"
+      )
+    ),
+    dialog.getByRole("button", { name: "Create Config" }).click(),
+  ]);
+
+  await expect(dialog).not.toBeVisible();
+  await expect(page.getByRole("cell", { name: configName })).toBeVisible();
+}
+
+async function createSecretKey(page: Page, key: string, value: string) {
+  await page.goto("/settings/secrets");
+  await page.waitForURL("**/settings/secrets");
+  await page.getByRole("button", { name: "New Secret" }).click();
+  const dialog = page.getByTestId("dialog");
+  await expect(dialog).toBeVisible();
+  await page.getByRole("textbox", { name: "Key" }).fill(key);
+  await page.getByLabel("Value").fill(value);
+  await Promise.all([
+    page.waitForResponse(
+      (resp) =>
+        resp.url().includes("/graphql") &&
+        resp.status() === 200 &&
+        (resp.request().postData()?.includes("SecretsMutationMutation") ??
+          false)
+    ),
+    page.getByRole("button", { name: "Create Secret" }).click(),
+  ]);
+  await expect(dialog).not.toBeVisible();
+}
+
 test.describe.serial("Code Evaluators", () => {
   const datasetName = `code-evals-${randomUUID().slice(0, 8)}`;
   const pythonSandboxName = `python-sandbox-${randomUUID().slice(0, 8)}`;
@@ -825,5 +883,123 @@ test.describe.serial("Code Evaluators", () => {
 
     await page.getByRole("button", { name: "Cancel" }).click();
     await expect(page.getByTestId("dialog")).not.toBeVisible();
+  });
+
+  // D6: Tab in the Python editor inserts four spaces, not a tab character.
+  test("Tab in Python code editor inserts four spaces", async ({ page }) => {
+    await gotoDatasetEvaluators(page, datasetName);
+
+    await page.getByRole("button", { name: "Add evaluator" }).click();
+    await page
+      .getByRole("menuitem", { name: "Create new code evaluator" })
+      .click();
+    const dialog = page.getByRole("dialog");
+    await expect(
+      page.getByRole("heading", { name: "Create Evaluator" })
+    ).toBeVisible();
+    await selectComboboxOption(page, "Sandbox", pythonSandboxName, dialog);
+
+    const editor = dialog.locator(".cm-content").first();
+    await editor.click();
+    await page.keyboard.press("ControlOrMeta+a");
+    await page.keyboard.press("Backspace");
+    await page.keyboard.press("Tab");
+
+    const lineText = await dialog
+      .locator(".cm-content .cm-line")
+      .first()
+      .textContent();
+    expect(lineText).toBe("    ");
+    expect(lineText).not.toContain("\t");
+
+    await page.getByRole("button", { name: "Cancel" }).click();
+    await expect(page.getByTestId("dialog")).not.toBeVisible();
+  });
+
+  // D4: literal env_var values are redacted from the evaluator details page.
+  test("redacts literal env_var values on evaluator details page", async ({
+    page,
+  }) => {
+    const sandboxName = `dogfood-redact-sandbox-${randomUUID().slice(0, 8)}`;
+    const evaluatorName = `dogfood-redact-eval-${randomUUID().slice(0, 8)}`;
+    const literalEnvVarName = "DOGFOOD_LITERAL_KEY";
+    const literalEnvVarValue = `super-secret-${randomUUID()}`;
+
+    await createE2BSandboxWithLiteralEnvVar(
+      page,
+      sandboxName,
+      literalEnvVarName,
+      literalEnvVarValue
+    );
+
+    await gotoDatasetEvaluators(page, datasetName);
+    await createCustomCodeEvaluator({
+      page,
+      evaluatorName,
+      language: "Python",
+      sandboxName,
+    });
+
+    await expectEvaluatorDetailsPage(page, evaluatorName);
+
+    const body = page.locator("body");
+    await expect(body).toContainText(literalEnvVarName);
+    await expect(body).toContainText("<redacted>");
+    await expect(body).not.toContainText(literalEnvVarValue);
+  });
+});
+
+// D5: env-var Name auto-populate behavior in the SandboxConfigDialog.
+// Independent from the eval/dataset fixtures above.
+test.describe
+  .serial("Code Evaluators — sandbox env-var Name auto-populate", () => {
+  const suffix = randomUUID().slice(0, 8).toUpperCase();
+  const secretKeyA = `DOGFOOD_AUTOPOP_A_${suffix}`;
+  const secretKeyB = `DOGFOOD_AUTOPOP_B_${suffix}`;
+
+  test("auto-populates Name on first selection and follows the previously selected key", async ({
+    page,
+  }) => {
+    await createSecretKey(page, secretKeyA, `value-a-${suffix}`);
+    await createSecretKey(page, secretKeyB, `value-b-${suffix}`);
+
+    await page.goto("/settings/sandboxes");
+    await page.waitForURL("**/settings/sandboxes");
+
+    await page.getByRole("button", { name: "New Sandbox" }).click();
+    const dialog = page.getByRole("dialog");
+    await expect(dialog).toBeVisible();
+
+    await dialog.getByLabel("Provider").fill("E2B");
+    await page.getByRole("option", { name: /E2B/i }).first().click();
+
+    await expect(dialog.getByText("Environment Variables")).toBeVisible();
+    await dialog.getByRole("button", { name: "Add Variable" }).click();
+
+    // Switch the env-var row to "Secret" kind.
+    await dialog.getByLabel("Kind").click();
+    await page.getByRole("option", { name: "Secret", exact: true }).click();
+
+    // The env-var Name input is the last "Name" input in the dialog;
+    // the first one is the sandbox config name.
+    const envVarNameInput = dialog.getByLabel("Name").last();
+    await expect(envVarNameInput).toHaveValue("");
+
+    // Blank Name → selecting a secret auto-populates Name with the key.
+    await selectComboboxOption(page, "Secret Key", secretKeyA, dialog);
+    await expect(envVarNameInput).toHaveValue(secretKeyA);
+
+    // Name still equals the previously selected key → switching secrets
+    // follows the new key.
+    await selectComboboxOption(page, "Secret Key", secretKeyB, dialog);
+    await expect(envVarNameInput).toHaveValue(secretKeyB);
+
+    // User-edited Name is preserved on subsequent secret changes.
+    await envVarNameInput.fill("CUSTOM_NAME");
+    await selectComboboxOption(page, "Secret Key", secretKeyA, dialog);
+    await expect(envVarNameInput).toHaveValue("CUSTOM_NAME");
+
+    await dialog.getByRole("button", { name: /cancel/i }).click();
+    await expect(dialog).not.toBeVisible();
   });
 });
